refac: Extract shared fixtures
