+++ tests/conftest.py
@@ -1,0 +1,36 @@
+"""Shared fixtures for integration tests."""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+import pytest
+
+
+def _find_repo_root() -> Path:
+    """Find the repository root by looking for .git directory."""
+    path = Path(__file__).resolve()
+    for parent in path.parents:
+        if (parent / ".git").exists():
+            return parent
+    raise RuntimeError("Could not find repository root")
+
+
+REPO_ROOT = _find_repo_root()
+
+
+@pytest.fixture
+def fake_home(tmp_path: Path) -> Path:
+    """Provide a fake HOME directory for test isolation.
+
+    Creates shell config files for testing profile modifications.
+    """
+    home = tmp_path / "home"
+    home.mkdir()
+    # Create shell config files
+    (home / ".bashrc").touch()
+    (home / ".zshrc").touch()
+    fish_config = home / ".config" / "fish"
+    fish_config.mkdir(parents=True)
+    (fish_config / "config.fish").touch()
+    return home
